requirements.txt
# $ pip install -r requirements.txt

bcrypt~=4.0.1
fastapi~=0.95.1
mysql-connector-python~=8.0.33
pydantic~=1.10.7
python-dateutil~=2.8.2
python-dotenv~=0.20.0
SQLAlchemy~=2.0.13
